Save the Run configurations for varios versions of IDE
Anybody should be able to easily run a test IDEA on a handful of predefined versions without any manual configuration.

To enable this a Gradle build had to be update to externalize a list of plugin dependencies, since newever release of IntelliJ require explicit java depency but old do not recognize it.